#57 Added schema checks in tests
